Revert "Lndl drafts"
