+++ lionagi/core/agent/base_agent.py
@@ -9,7 +9,7 @@ from lionagi.libs import func_call, AsyncUtil
 
 
 from lionagi.core.mail.schema import StartMail
-from lionagi.core.generic.node import Node
+from lionagi.core.generic import Node
 from lionagi.core.mail.mail_manager import MailManager
 from lionagi.core.execute.base_executor import BaseExecutor
 from lionagi.core.execute.structure_executor import StructureExecutor
