--- lionagi/core/generic/model.py
@@ -1,5 +1,0 @@
-from abc import ABC
-
-from .component import BaseComponent
-
-
