+++ lionagi/experimental/report/report.py
@@ -2,7 +2,7 @@ from typing import Any, Type
 from pydantic import Field
 
 # from lionagi import logging as _logging
-from lionagi.core.generic.component import BaseComponent
+from lionagi.core.generic import BaseComponent
 from lionagi.experimental.report.form import Form
 from lionagi.experimental.report.util import get_input_output_fields
