+++ lionagi/experimental/work/work_function.py
