--- lionagi/experimental/work2/work/__init__.py
@@ -1,10 +1,0 @@
-from .schema import WorkLog
-from .work_function import WorkFunction, work
-from .worker import Worker
-
-__all__ = [
-    "WorkLog",
-    "WorkFunction",
-    "work",
-    "Worker",
-]
