+++ lionagi/integrations/chunker/chunk.py
@@ -1,7 +1,7 @@
 from typing import Union, Callable
 
 from lionagi.libs import func_call
-from lionagi.core.generic.node import Node
+from lionagi.core.generic import Node
 from ..bridge.langchain_.langchain_bridge import LangchainBridge
 from ..bridge.llamaindex_.llama_index_bridge import LlamaIndexBridge
