+++ lionagi/libs/ln_func_call.py
@@ -294,7 +294,6 @@ async def rcall(
     backoff_factor: float = 2.0,
     default: Any = None,
     timeout: float | None = None,
-    timing: bool = False,
     **kwargs,
 ) -> Any:
     """
@@ -339,13 +338,10 @@ async def rcall(
     """
     last_exception = None
     result = None
-    start_time = SysUtil.get_now(datetime_=False)
+
     for attempt in range(retries + 1) if retries == 0 else range(retries):
         try:
-            result = await _tcall(func, *args, timeout=timeout, **kwargs)
-            if timing:
-                return (result, SysUtil.get_now(datetime_=False) - start_time)
-            return result
+            return await _tcall(func, *args, timeout=timeout, **kwargs)
         except Exception as e:
             last_exception = e
             if attempt < retries:
@@ -1097,14 +1093,6 @@ class CallDecorator:
 
         return wrapper
 
-    @staticmethod
-    def count_calls(func):
-        async def wrapper(*args, **kwargs):
-            wrapper.calls += 1
-            return await func(*args, **kwargs)
-
-        wrapper.calls = 0
-        return wrapper
 
 class Throttle:
     """
